Pre-Alpha Version 0.4
Pre-Alpha Version 0.3 - Slight restructuring of code for better readability and ease of understanding. We have created the first enemy, and created a system where it will be easy to create as many enemies as we need without needing to create a table for each one. In the next update, we will be able to give interaction between the player and the enemy (i.e: taking damage)