Pre-Alpha Version 0.5
Pre-Alpha Version 0.5 - Doors have been created. As of now, enemies are unable to go through them, and the player is able to go through them. The first area has also been mapped out, and more sprites have been created for the first part of the map. Ceiling collision will be implemented in the next update, as well as the ability to actually destroy the doors (and enemies) instead of simply going through them.