Pre-Alpha Version 0.5.2
Pre-Alpha Version 0.5.2 - Added a more dynamic camera, and have the intro to HK more replicated with the falling. Camera is limited in such a way that you won't be able to see "out of bounds" below the screen, and future updates will implement it such that you cannot look past the false wall until you've walked into it on the left, and making a border.