Pre-Alpha 0.6
Pre-Alpha 0.6 - Violence has been achieved, the player is now able to take damage, and likewise enemies can be damaged. In a future update, doors will also be able to be destroyed, while also making enemies and doors have actual physics with the player and each other. Restructuring of code has occurred again to try and make it easier to read and understand.