Pre-Alpha 0.7
Pre-Alpha 0.7 - Doors and Enemies now have proper collision with the player, which will allow enemies to dynamically have a larger movement radius, and forces the player to break doors. Future updates will further finetune the enemy collision while also creating knockback for when the player is hit by an enemy, or if the player hits an enemy with their nail.