Pre-Alpha 0.7.5
Pre-Alpha 0.7.5 - Updated the formula for collision for players and enemies, still needs a good bit of work. Reworked a few of the functions involving that. I plan to spend more time polishing it, as well as how "being hit" works. We're not far off from a game here. Next goal should be to do a massive overhaul for the code, it's a bit of a mess and could use some polish.

GOALS 
- Consolidate all of the Collision functions into one collision
- Remove redundant code
- Make it easy to read and understand