Pre-Alpha 0.7.7
Pre-Alpha 0.7.7 - We've revamped the damage system and created two visual marks for when the player takes damage, a white "damage" particle around the player, as well as a fading away close-in on the screen. With these two things, it should be more clear when the player is hurt, on top of losing health and the screen freezing. Finally, the player will now properly be sent flying away in either direction when they take damage. Next patch should hopefully continue making code easier to read and understand, to be less spaghetti.